# Optimize Request.builder().from()