--- a/dialogue-jmh/src/main/java/com/palantir/dialogue/core/RequestBuilderBenchmark.java
+++ b/dialogue-jmh/src/main/java/com/palantir/dialogue/core/RequestBuilderBenchmark.java
@@ -28,7 +28,6 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.Threads;
 import org.openjdk.jmh.annotations.Warmup;
-import org.openjdk.jmh.profile.GCProfiler;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
@@ -82,7 +81,7 @@ public class RequestBuilderBenchmark {
                 .include(RequestBuilderBenchmark.class.getSimpleName())
                 .jvmArgsPrepend("-Xmx1024m", "-Xms1024m", "-XX:+CrashOnOutOfMemoryError")
                 // .jvmArgsPrepend("-XX:+FlightRecorder", "-XX:StartFlightRecording=filename=./foo.jfr")
-                .addProfiler(GCProfiler.class)
+                // .addProfiler(GCProfiler.class)
                 .build();
         new Runner(opt).run();
     }