--- a/dialogue-target/src/main/java/com/palantir/dialogue/Request.java
+++ b/dialogue-target/src/main/java/com/palantir/dialogue/Request.java
@@ -31,7 +31,7 @@ import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
 
-/** Defines the parameters of a single {@link Call} to an {@link Endpoint}. */
+/** Defines the parameters of a single call to an {@link Endpoint}. */
 @ThreadSafe
 public final class Request {
 
@@ -44,7 +44,7 @@ public final class Request {
         body = builder.body;
         headerParams = builder.unmodifiableHeaderParams();
         queryParams = builder.unmodifiableQueryParams();
-        pathParams = builder.pathParams.build();
+        pathParams = builder.unmodifiablePathParams();
     }
 
     /**
@@ -128,17 +128,22 @@ public final class Request {
         @Nullable
         private ImmutableListMultimap.Builder<String, String> queryParams;
 
-        private ImmutableMap.Builder<String, String> pathParams = ImmutableMap.builder();
+        @Nullable
+        private ImmutableMap.Builder<String, String> pathParams;
 
         private Optional<RequestBody> body = Optional.empty();
 
-        // To optimize the case where we don't end up modifying headers/queryparams/pathparams, we may store a
+        // To optimize the case where a builder doesn't end up modifying headers/queryparams/pathparams, we may store
         // references to another Request's internal objects. If we need to do a mutation, then we'll copy the elements
+        // and null these out.
         @Nullable
         private ListMultimap<String, String> existingUnmodifiableHeaderParams;
 
         @Nullable
         private ImmutableListMultimap<String, String> existingQueryParams;
+
+        @Nullable
+        private ImmutableMap<String, String> existingPathParams;
 
         private Builder() {}
 
@@ -147,62 +152,13 @@ public final class Request {
 
             existingUnmodifiableHeaderParams = Multimaps.unmodifiableListMultimap(existing.headerParams);
             existingQueryParams = existing.queryParams;
-            putAllPathParams(existing.pathParams());
+            existingPathParams = existing.pathParams;
+
             Optional<RequestBody> bodyOptional = existing.body();
             if (bodyOptional.isPresent()) {
                 body(bodyOptional);
             }
             return this;
-        }
-
-        private ListMultimap<String, String> mutableHeaderParams() {
-            if (headerParams == null) {
-                headerParams = MultimapBuilder.treeKeys(String.CASE_INSENSITIVE_ORDER)
-                        .arrayListValues()
-                        .build();
-
-                if (existingUnmodifiableHeaderParams != null) {
-                    headerParams.putAll(existingUnmodifiableHeaderParams);
-                    existingUnmodifiableHeaderParams = null;
-                }
-            }
-            return headerParams;
-        }
-
-        private ImmutableListMultimap.Builder<String, String> mutableQueryParams() {
-            if (queryParams == null) {
-                queryParams = ImmutableListMultimap.builder();
-
-                if (existingQueryParams != null) {
-                    queryParams.putAll(existingQueryParams);
-                    existingQueryParams = null;
-                }
-            }
-            return queryParams;
-        }
-
-        private ListMultimap<String, String> unmodifiableHeaderParams() {
-            if (existingUnmodifiableHeaderParams != null) {
-                return existingUnmodifiableHeaderParams;
-            }
-
-            if (headerParams != null) {
-                return Multimaps.unmodifiableListMultimap(headerParams);
-            }
-
-            return EMPTY;
-        }
-
-        public ImmutableListMultimap<String, String> unmodifiableQueryParams() {
-            if (existingQueryParams != null) {
-                return existingQueryParams;
-            }
-
-            if (queryParams != null) {
-                return queryParams.build();
-            }
-
-            return EMPTY;
         }
 
         public Request.Builder putHeaderParams(String key, String... values) {
@@ -260,22 +216,22 @@ public final class Request {
         }
 
         public Request.Builder putPathParams(String key, String value) {
-            pathParams.put(key, value);
+            mutablePathParams().put(key, value);
             return this;
         }
 
         public Request.Builder putPathParams(Map.Entry<String, ? extends String> entry) {
-            pathParams.put(entry);
+            mutablePathParams().put(entry);
             return this;
         }
 
         public Request.Builder pathParams(Map<String, ? extends String> entries) {
-            pathParams = ImmutableMap.builder();
+            pathParams = null;
             return putAllPathParams(entries);
         }
 
         public Request.Builder putAllPathParams(Map<String, ? extends String> entries) {
-            pathParams.putAll(entries);
+            mutablePathParams().putAll(entries);
             return this;
         }
 
@@ -288,6 +244,80 @@ public final class Request {
         public Request.Builder body(Optional<? extends RequestBody> value) {
             body = (Optional<RequestBody>) value;
             return this;
+        }
+
+        private ListMultimap<String, String> mutableHeaderParams() {
+            if (headerParams == null) {
+                headerParams = MultimapBuilder.treeKeys(String.CASE_INSENSITIVE_ORDER)
+                        .arrayListValues()
+                        .build();
+
+                if (existingUnmodifiableHeaderParams != null) {
+                    headerParams.putAll(existingUnmodifiableHeaderParams);
+                    existingUnmodifiableHeaderParams = null;
+                }
+            }
+            return headerParams;
+        }
+
+        private ImmutableListMultimap.Builder<String, String> mutableQueryParams() {
+            if (queryParams == null) {
+                queryParams = ImmutableListMultimap.builder();
+
+                if (existingQueryParams != null) {
+                    queryParams.putAll(existingQueryParams);
+                    existingQueryParams = null;
+                }
+            }
+            return queryParams;
+        }
+
+        private ImmutableMap.Builder<String, String> mutablePathParams() {
+            if (pathParams == null) {
+                pathParams = ImmutableMap.builder();
+
+                if (existingPathParams != null) {
+                    pathParams.putAll(existingPathParams);
+                    existingPathParams = null;
+                }
+            }
+            return pathParams;
+        }
+
+        private ListMultimap<String, String> unmodifiableHeaderParams() {
+            if (existingUnmodifiableHeaderParams != null) {
+                return existingUnmodifiableHeaderParams;
+            }
+
+            if (headerParams != null) {
+                return Multimaps.unmodifiableListMultimap(headerParams);
+            }
+
+            return EMPTY;
+        }
+
+        private ImmutableListMultimap<String, String> unmodifiableQueryParams() {
+            if (existingQueryParams != null) {
+                return existingQueryParams;
+            }
+
+            if (queryParams != null) {
+                return queryParams.build();
+            }
+
+            return EMPTY;
+        }
+
+        private ImmutableMap<String, String> unmodifiablePathParams() {
+            if (existingPathParams != null) {
+                return existingPathParams;
+            }
+
+            if (pathParams != null) {
+                return pathParams.build();
+            }
+
+            return ImmutableMap.of();
         }
 
         public Request build() {