## Before this PR

As per a discussion in #dev-foundry-infra, a decent proportion of dialogue overhead is actually spent copying Requests just to add/modify a header. 

Rather than trying to optimize multimaps, we can actually get a bit of a speedup by just re-using stuff where possible. (There's no need to allocate and populate an identical map to an existing one!)

```
Benchmark                                                                     Mode  Cnt     Score     Error   Units
RequestBuilderBenchmark.addUserAgentToEmpty                                  thrpt   12  9310.828 ± 736.043  ops/ms
RequestBuilderBenchmark.addUserAgentToNonEmpty                               thrpt   12   936.356 ±  71.790  ops/ms

RequestBuilderBenchmark.addUserAgentToEmpty:·gc.alloc.rate.norm              thrpt   12   592.000 ±   0.001    B/op
RequestBuilderBenchmark.addUserAgentToNonEmpty:·gc.alloc.rate.norm           thrpt   12  2688.003 ±   0.008    B/op
```


## After this PR
==COMMIT_MSG==
Reduce allocation and eliminate some unnecessary copying from `Request.builder().from()`
==COMMIT_MSG==

Less GC due to dumb copying means servers will have more CPU available to do real work.

```
Benchmark                                                                     Mode  Cnt      Score      Error   Units
RequestBuilderBenchmark.addUserAgentToEmpty                                  thrpt   12  23070.647 ± 3011.135  ops/ms
RequestBuilderBenchmark.addUserAgentToNonEmpty                               thrpt   12   2175.030 ±  212.811  ops/ms

RequestBuilderBenchmark.addUserAgentToEmpty:·gc.alloc.rate.norm              thrpt   12    264.000 ±    0.001    B/op
RequestBuilderBenchmark.addUserAgentToNonEmpty:·gc.alloc.rate.norm           thrpt   12    744.001 ±    0.004    B/op
```

## Possible downsides?

- the _vast_ majority of dialogue users will not care about this, it's really only helpful in the case where something is making tons of network requests and is pretty maxed out on CPU...
- whenever we're micro-optimizing, there's a danger that making the code harder to understand is actually worse than a small performance hit.  I _hope_ this implementation is still easy to understand, but would like to hear if not.