@svc-autorelease 2020-05-26
Released [1.51.0](https://github.com/palantir/dialogue/releases/tag/1.51.0)